requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
cymem==2.0.6
datasets==2.4.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Django==4.1.1
entrypoints==0.4
executing==0.9.1
fastai==2.7.7
fastbook==0.0.26
fastcore==1.5.11
fastdownload==0.0.7
fastjsonschema==2.16.1
fastprogress==1.0.3
filelock==3.7.1
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.7.1
graphviz==0.20.1
huggingface-hub==0.8.1
idna==3.3
image==1.5.33
ipydatawidgets==4.3.1.post1
ipykernel==6.15.1
ipympl==0.9.2
ipysegment==0.1.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.10.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kaggle==1.5.12
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mpl-interactions==0.22.0
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
responses==0.18.0
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.12.0
Send2Trash==1.8.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.4
stack-data==0.3.0
terminado==0.15.0
text-unidecode==1.3
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torchaudio==0.12.0+cu116
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
traittypes==0.2.1
transformers==4.21.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.7.2
